rename: 게임 페이지 관련 폴더 구조 변경
